--- app/src/main/java/com/example/pinkpanterwear/ui/UserCategoryViewModel.kt
@@ -1,42 +1,0 @@
-package com.example.pinkpanterwear.ui
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.example.pinkpanterwear.data.ProductRepository
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.launch
-
-class UserCategoryViewModel : ViewModel() {
-
-    private val repository = ProductRepository() // TODO: Use dependency injection
-
-    private val _categories = MutableStateFlow<List<String>>(emptyList())
-    val categories: StateFlow<List<String>> = _categories.asStateFlow()
-
-    private val _isLoading = MutableStateFlow(false)
-    val isLoading: StateFlow<Boolean> = _isLoading.asStateFlow()
-
-    private val _error = MutableStateFlow<String?>(null)
-    val error: StateFlow<String?> = _error.asStateFlow()
-
-    init {
-        fetchCategories()
-    }
-
-    fun fetchCategories() {
-        viewModelScope.launch {
-            _isLoading.value = true
-            _error.value = null
-            try {
-                _categories.value = repository.getAllCategoriesFromFirestore()
-            } catch (e: Exception) {
-                _error.value = "Failed to fetch categories: ${e.message}"
-                // Log.e("UserCategoryViewModel", "Error fetching categories", e)
-            } finally {
-                _isLoading.value = false
-            }
-        }
-    }
-}
